+++ src/index.js
@@ -1,3 +1,4 @@
 export { default as PersianDateTimePicker } from './PersianDateTimePicker.jsx';
 export { default as PersianDateRangePicker } from './PersianDateRangePicker.jsx';
 export { default as PersianCalendar } from './PersianCalendar.jsx';
+export { default as PersianTimePicker } from './PersianTimePicker.jsx';